[ex4][part1] feat: Implement peek method in List1 module
This commit adds the implementation of the `peek` method in the `List1` module. The `peek` method returns a reference to the first element of the list. It matches the `head` field of the list and returns `None` if it is `ListLink::Nil`, otherwise it returns `Some(element)` where `element` is the first element of the list.